Fix very minor spelling issue (build → built)
Fixed a very minor spelling issue I happened to notice. "Build" is used
for the present tense ("I will build this"), and "built" is for the past
tense ("I built this").

It's one of those confusing language things 😅 The same applies to some
other verbs ending with a "d" by the way (send probably being the most
common).